[#47] refactor : 로그인, 회원탈퇴 컴포넌트로 분리
